fix: validation fails because of status field wrongly set
